+++ processing/providers.py
@@ -1,24 +1,51 @@
 # See https://docs.qgis.org/3.10/en/docs/pyqgis_developer_cookbook/processing.html
 from qgis.core import QgsProcessingProvider
 from qgis.PyQt.QtGui import QIcon
-from threedi_results_analysis.processing.dwf_calculation_algorithm import DWFCalculatorAlgorithm
-from threedi_results_analysis.processing.gpkg_conversion_algorithm import ThreeDiConvertToGpkgAlgorithm
-from threedi_results_analysis.processing.grid_creation_algorithm import ThreeDiGenerateCompGridAlgorithm
-from threedi_results_analysis.processing.cross_sectional_discharge_algorithm import CrossSectionalDischargeAlgorithm
+from threedi_results_analysis.processing.cross_sectional_discharge_algorithm import (
+    CrossSectionalDischargeAlgorithm,
+)
+from threedi_results_analysis.processing.dwf_calculation_algorithm import (
+    DWFCalculatorAlgorithm,
+)
+from threedi_results_analysis.processing.gpkg_conversion_algorithm import (
+    ThreeDiConvertToGpkgAlgorithm,
+)
+from threedi_results_analysis.processing.grid_creation_algorithm import (
+    ThreeDiGenerateCompGridAlgorithm,
+)
 from threedi_results_analysis.processing.leak_detector_algorithms import (
     DetectLeakingObstaclesAlgorithm,
+)
+from threedi_results_analysis.processing.leak_detector_algorithms import (
     DetectLeakingObstaclesWithDischargeThresholdAlgorithm,
+)
+from threedi_results_analysis.processing.rasters_to_netcdf_algorithm import (
+    RastersToNetCDFAlgorithm,
 )
 from threedi_results_analysis.processing.schematisation_algorithms import (
     CheckSchematisationAlgorithm,
-    MigrateAlgorithm,
-    ImportSufHydAlgorithm,
+)
+from threedi_results_analysis.processing.schematisation_algorithms import (
     GuessIndicatorAlgorithm,
+)
+from threedi_results_analysis.processing.schematisation_algorithms import (
     ImportHydXAlgorithm,
 )
-from threedi_results_analysis.processing.rasters_to_netcdf_algorithm import RastersToNetCDFAlgorithm
-
-from threedi_results_analysis.processing.threedidepth_algorithms import ThreediDepthAlgorithm, ThreediMaxDepthAlgorithm
+from threedi_results_analysis.processing.schematisation_algorithms import (
+    ImportSufHydAlgorithm,
+)
+from threedi_results_analysis.processing.schematisation_algorithms import (
+    MigrateAlgorithm,
+)
+from threedi_results_analysis.processing.structure_control_action_algorithms import (
+    StructureControlActionAlgorithm,
+)
+from threedi_results_analysis.processing.threedidepth_algorithms import (
+    ThreediDepthAlgorithm,
+)
+from threedi_results_analysis.processing.threedidepth_algorithms import (
+    ThreediMaxDepthAlgorithm,
+)
 
 import os
 
@@ -41,6 +68,7 @@ class ThreediProvider(QgsProcessingProvider):
         self.addAlgorithm(DetectLeakingObstaclesAlgorithm())
         self.addAlgorithm(DetectLeakingObstaclesWithDischargeThresholdAlgorithm())
         self.addAlgorithm(RastersToNetCDFAlgorithm())
+        self.addAlgorithm(StructureControlActionAlgorithm())
 
     def id(self, *args, **kwargs):
         """The ID of your plugin, used for identifying the provider.